Server/requirements.txt
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.15.1
gunicorn==21.2.0
h11==0.14.0
packaging==24.0
pillow==10.3.0
psycopg-binary==3.1.18
psycopg2==2.9.9
uvicorn==0.29.0
djoser
djangorestframework_simplejwt